requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
httpx
gunicorn
